+++ lib/interfaces/FunctionOptions.ts
@@ -1,7 +1,11 @@
-import {PubSub} from "@google-cloud/pubsub";
+import { PubSub } from "@google-cloud/pubsub";
 export interface FunctionOptions {
-    functionIdentifier: string;
+    functionIdentifier?: string;
     errorCode?: string;
     pubSubClient?: PubSub;
     errorTopic?: string;
+    correlationId?: string;
+    projectId?: string;
+    apifsSecretHeader?: string;
+    apifsSecretValue?: string;
 }