+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # gcf-helper CHANGELOG
 
+## 2019-03-19, Version 0.2.0-0.3.0
+
+* reading default FunctionOptions from environment variables (see DOC.md)
+* auto handling of pubsub instance if possible
+* added `validateRequestAuthorization` method for http functions
+
 ## 2019-03-15, Version 0.1.0
 
 * Initial release with changelog